back/requirements.txt
# ===========================================
# Core Framework
# ===========================================
fastapi[standard]>=0.112.0
uvicorn>=0.23.0
starlette>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# ===========================================
# AI Agent Framework
# ===========================================
agno[openai]>=2.0.0
openai>=1.0.0
exa-py
ddgs
nest_asyncio
google-search-results
beautifulsoup4
lxml

# ===========================================
# Database
# ===========================================
sqlalchemy>=2.0.0
peewee>=3.15.0
psycopg2-binary>=2.9.0
asyncpg>=0.28.0

# ===========================================
# HTTP Clients
# ===========================================
httpx>=0.24.0
requests>=2.28.0
aiohttp>=3.8.0

# ===========================================
# Data Processing
# ===========================================
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# ===========================================
# Technical Analysis
# ===========================================
ta>=0.10.0
pandas-ta

# ===========================================
# Crypto & Trading
# ===========================================
ccxt>=4.0.0
yfinance>=0.2.0

# ===========================================
# Utilities
# ===========================================
schedule>=1.2.0
pydantic[email]>=2.0.0
python-jose>=3.3.0
PyJWT>=2.8.0
email-validator>=2.0.0
markdown>=3.4.0

# ===========================================
# Alibaba Cloud (for OSS & Email)
# ===========================================
oss2>=2.17.0
aliyun-python-sdk-core>=2.13.0

# ===========================================
# Security & Encryption
# ===========================================
cryptography>=41.0.0
